[#40]  Refactored code
