+++ lib/api/repository/survey_repository.dart
@@ -1,3 +1,4 @@
+import 'package:flutter_survey/api/request/submit_survey_question_request.dart';
 import 'package:flutter_survey/api/request/submit_survey_request.dart';
 import 'package:flutter_survey/api/survey_service.dart';
 import 'package:flutter_survey/api/exception/network_exceptions.dart';
